+++ test/typescript/apis/project/cds-ql.ts
@@ -4,6 +4,13 @@ import { Foo, Foos, attach } from './dummy'
 // unwrapped plural types
 let sel: SELECT<Foo>
 sel = SELECT(Foo)
+sel = SELECT(Foo, 42)
+sel = SELECT(Foo.drafts)
+sel = SELECT(Foos.drafts)
+sel = SELECT.from(Foo.drafts)
+sel = SELECT.from(Foos.drafts)
+sel = SELECT.from(Foos.drafts, 42)
+
 const selStatic: SELECT<Foos> | Promise<Foos> = SELECT.from(Foos)
 
 SELECT.from(Foos).columns("x") // x was suggested by code completion
@@ -19,6 +26,8 @@ s.SELECT.from.ref
 s.SELECT.columns?.[0].ref
 s.SELECT.where?.[0].ref
 s.SELECT.where?.[2].val
+
+SELECT(Foos) === SELECT.from(Foos)
 
 INSERT.into(Foos).columns("x") // x was suggested by code completion
 let ins: INSERT<Foo>